[FE] rename: slice 이름 변경
- CreateGroupInfo -> CreateGroupData
- issue (#10)
