+++ client/src/pages/GroupCreatePage/DateInput/AntDatePicker.tsx
@@ -8,11 +8,11 @@ import 'antd/es/date-picker/style/css';
 import { useDispatch, useSelector } from 'react-redux';
 import { ReducerType } from '../../../store/rootReducer';
 import { GroupData } from '../../../types/CreateGroup';
-import { setGroupData } from '../../../store/slices/createGroupInfo';
+import { setGroupData } from '../../../store/slices/createGroupData';
 
 function AntDatePicker(): JSX.Element {
   const { startDate, endDate } = useSelector<ReducerType, GroupData>(
-    (state) => state.createGroupInfo,
+    (state) => state.createGroupData,
   );
   const dispatch = useDispatch();
